Added support to 64 bits file lenght. This make it possible to create storages bigger than 2 GB

+++ src/datastorage.h
@@ -49,7 +49,7 @@ typedef struct {
 	rw_lock_t lock;
 }datastore_t;
 
-datastore_t* datastore_create(char* name, int storage_size, int index_length);
+datastore_t* datastore_create(char* name, uint64_t storage_size, uint64_t index_length);
 
 char* datastore_lookup(datastore_t* datastore, char* key);
 
@@ -59,11 +59,11 @@ int datastore_set(datastore_t* datastore, char* key, char* value);
 
 int datastore_remove(datastore_t* datastore, char* key);
 
-int datastore_keys_number(datastore_t* datastore);
+uint64_t datastore_keys_number(datastore_t* datastore);
 
-int datastore_count_keys(datastore_t* datastore);
+uint64_t datastore_count_keys(datastore_t* datastore);
 
-void datastore_list_keys(datastore_t* datastore, char **keys, int len);
+void datastore_list_keys(datastore_t* datastore, char **keys, uint64_t len);
 
 void datastore_clear(datastore_t* datastore);
